backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# Validation
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Data processing
numpy==1.26.2
pandas==2.1.4

# Background tasks (optional)
celery==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
